# apps/mrerp/backend/openapi.yaml
openapi: 3.0.3
info:
  title: MRERP API
  version: 1.0.0
  description: MRERP modular-monolith API contract
paths:
...
  /api/v1/leave/attendance-imports/:
    get:
      operationId: api_v1_leave_attendance_imports_list
      tags:
      - api
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceImportResult'
          description: ''
  /api/v1/leave/attendance-imports/{id}/commit/:
    post:
      operationId: api_v1_leave_attendance_imports_commit_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Commit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Commit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Commit'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceImportResult'
          description: ''
  /api/v1/leave/attendance-imports/preview/:
    post:
      operationId: api_v1_leave_attendance_imports_preview_create
      tags:
      - api
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Upload'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceImportResult'
          description: ''
...
components:
  schemas:
...
    AttendanceImportResult:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        filename:
          type: string
        month:
          type: string
        created_at:
          type: string
          format: date-time
        committed_at:
          type: string
          format: date-time
          nullable: true
        row_count:
          type: integer
        replaced_count:
          type: integer
        imported_by:
          type: string
        source: {}
        mappings:
          type: object
          additionalProperties: {}
        employee_options:
          type: array
          items:
            type: object
            additionalProperties: {}
        existing_days:
          type: object
          additionalProperties: {}
      required:
      - committed_at
      - created_at
      - filename
      - imported_by
      - month
      - replaced_count
      - row_count
      - uuid
...
    Commit:
      type: object
      properties:
        mappings:
          type: object
          additionalProperties:
            type: string
            format: uuid
        replace_existing:
          type: boolean
          default: false
      required:
      - mappings
...
    Upload:
      type: object
      properties:
        file:
          type: string
          format: uri
      required:
      - file
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid